python-backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
PyMuPDF>=1.23.8
ollama>=0.1.7
numpy>=1.24.3
pandas>=2.1.4
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.2
loguru>=0.7.2
tenacity>=8.2.3
tqdm>=4.66.1
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0